pj3_group_work/requirements.txt
einops==0.8.0
ftfy==6.2.0
h5py==3.10.0
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
regex==2024.5.15
scipy==1.13.0
setuptools==65.5.0
tensorboardX==2.6.2.2
timm==1.0.3
torch==2.2.2+cu121
torchvision==0.17.2
tqdm==4.66.2
